embedchain-backend/requirements.txt
python-dotenv
embedchain
fastapi==0.108.0
uvicorn==0.25.0
beautifulsoup4
transformers
sentence-transformers
torch
langchain_huggingface
pydantic==2.4.2
python-multipart==0.0.6